requirements.txt
africastalking>=1.2.8,<2.0.0
flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<6.0.0
gunicorn>=22.0.0,<24.0.0
joblib>=1.3.0,<2.0.0
numpy>=1.24.0,<3.0.0
pandas>=2.0.0,<3.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
scikit-learn>=1.2.0,<2.0.0
spacy>=3.6.0,<4.0.0
sqlalchemy>=2.0.0,<3.0.0
twilio>=8.0.0,<10.0.0
